修改检查资源引用合法性的方法中的判断，修改为使用int值进行比较

+++ yaya/jcms/src/main/java/cn/medcn/jcms/controller/MeetController.java
@@ -338,7 +338,7 @@ public class MeetController extends BaseController {
             }
             Principal principal = SubjectUtils.getCurrentUser();
             //判断资源是否属于当前用户
-            if (principal.getId() != course.getOwner()) {
+            if (principal.getId().intValue() != course.getOwner().intValue()) {
                 throw new SystemException("您无法引用不属于您的资源");
             }
         }
